fix get all post, fix del post, fix get all media files, add media-files/remove/id endpoint
- fix get all post when db length <=0
- fix del post by id, only your post now can be deleted
- fix get all media files when db length <=0
- add media files remove endpoint

+++ controllers/ownPosts/removeOwnPost.js
@@ -4,7 +4,15 @@ const { HttpError } = require("../../routes/errors/HttpErrors");
 const { ownPostTransformer } = require("../../helpers/index");
 
 const removeOwnPost = async (req, res, next) => {
+  const { _id } = req.user;
   const { postId } = req.params;
+
+  const post = await Post.findById({ _id: postId });
+
+  if ((await post.owner.toString()) !== _id.toString()) {
+    throw HttpError(404, "Not found");
+  }
+
   const result = await Post.findByIdAndDelete({ _id: postId });
 
   if (!result) {
